fix: update u16::MAX to u32::MAX in vault tests to match function signature

+++ contracts/vault/src/test_reentrancy.rs
@@ -41,7 +41,7 @@ impl MaliciousToken {
                 let client = CalloraVaultClient::new(&env, &vault);
 
                 // Attempt re-entry into deduct
-                let _ = client.try_deduct(&caller, &1, &Some(Symbol::new(&env, "reentry_token")), &u16::MAX);
+                let _ = client.try_deduct(&caller, &1, &Some(Symbol::new(&env, "reentry_token")), &u32::MAX);
             }
         }
     }
@@ -103,7 +103,7 @@ impl MaliciousSettlement {
                 let client = CalloraVaultClient::new(&env, &vault);
 
                 // Attempt re-entry into deduct
-                let _ = client.try_deduct(&caller, &1, &Some(Symbol::new(&env, "reentry_settle")), &u16::MAX);
+                let _ = client.try_deduct(&caller, &1, &Some(Symbol::new(&env, "reentry_settle")), &u32::MAX);
             }
         }
     }
@@ -155,7 +155,7 @@ fn test_reentrancy_via_token_transfer_is_blocked_by_auth() {
     assert_eq!(initial_balance, 1000);
 
     // Trigger deduct -> calls token.transfer -> calls vault.deduct (re-entry)
-    let result = vault_client.try_deduct(&owner, &100, &Some(Symbol::new(&env, "first_call")), &u16::MAX);
+    let result = vault_client.try_deduct(&owner, &100, &Some(Symbol::new(&env, "first_call")), &u32::MAX);
 
     assert!(result.is_ok(), "First deduct should succeed");
     assert_eq!(
@@ -197,7 +197,7 @@ fn test_reentrancy_via_settlement_callback_is_blocked() {
     assert_eq!(initial_balance, 1000);
 
     // Trigger deduct -> calls settlement.receive_payment -> calls vault.deduct (re-entry)
-    let result = vault_client.try_deduct(&owner, &100, &Some(Symbol::new(&env, "first_call")), &u16::MAX);
+    let result = vault_client.try_deduct(&owner, &100, &Some(Symbol::new(&env, "first_call")), &u32::MAX);
 
     assert!(result.is_ok(), "First deduct should succeed");
     assert_eq!(
@@ -298,7 +298,7 @@ fn test_reentrancy_by_authorized_attacker() {
     assert_eq!(initial_balance, 1000);
 
     // Attacker calls deduct -> token.transfer -> attacker calls vault.deduct (re-entry)
-    let result = vault_client.try_deduct(&attacker, &100, &Some(Symbol::new(&env, "first_call")), &u16::MAX);
+    let result = vault_client.try_deduct(&attacker, &100, &Some(Symbol::new(&env, "first_call")), &u32::MAX);
 
     assert!(result.is_ok(), "First deduct should succeed");
     assert_eq!(
